impl 0.01 STEP 1

+++ src/XYG_BASIC/HelpAA.java
@@ -94,7 +94,7 @@ public class HelpAA {
                 point = new MyPoint(val, val);
             } else {
                 double val = (Double) value;
-                point = new MyPoint((int) val, val);
+                point = new MyPoint(val, val);
             }
         } else {
             point = (MyPoint) value;
